# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NotedThat HTTP API",
    "description": "The versioned machine API of a NotedThat server. Paths are relative to the server URL, `/api/v1`.\n\nThis document is generated from the server's code and is authoritative for paths, parameters, schemas and status codes. `docs/API.md` explains behaviour; where the two disagree, `docs/API.md` is wrong.\n\nEvery response carries `x-request-id`, echoing the request's when it sent one.",
    "contact": {
      "name": "NotedThat Contributors"
    },
    "license": {
      "name": "MPL-2.0",
      "identifier": "MPL-2.0"
    },
    "version": "0.12.3"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/knowledgebases/{kb_slug}/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Subscribe to a knowledge base's change events.",
        "description": "A server-sent event stream (`text/event-stream`). The first frame is\n`retry: 3000` and the comment `:subscribed`; each event after it is\n\n```text\nid: <position>\nevent: object.written | object.deleted | object.indexed | object.index_failed\ndata: <ObjectEvent as JSON>\n```\n\nwith a `keep-alive` comment every 15 seconds while idle. Only events for\nkeys the caller may `list` are sent. Reconnect with `Last-Event-ID` to\nresume after the last event received. See the `ObjectEvent` schema for the\n`data` payload. A server with events disabled answers `404`.",
        "operationId": "subscribe_events",
        "parameters": [
          {
            "name": "kb_slug",
            "in": "path",
            "description": "The knowledge base.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KbSlug"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "description": "Only events for keys starting with this prefix.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "Only this kind: `written`, `deleted`, `indexed` or `index_failed`. An\n`object.` prefix is accepted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mime",
            "in": "query",
            "description": "Only events carrying this content type, exactly or as `type/*`. A\ndeletion carries none, so never matches.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "description": "Resume after this position: the `id` of the last event received.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event stream. Each `data` line is an `ObjectEvent`.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`"
              },
              "X-Accel-Buffering": {
                "schema": {
                  "type": "string"
                },
                "description": "`no`"
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          },
          "503": {
            "$ref": "#/components/responses/BackendUnavailable"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "The error envelope every JSON error response carries.",
        "required": [
          "error",
          "message",
          "request_id"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": "string",
            "description": "The request's id, the same value as the `x-request-id` response header."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable error code. Clients branch on this, never on `message`.",
        "enum": [
          "invalid_request",
          "malformed_range",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "gone",
          "payload_too_large",
          "precondition_failed",
          "no_match",
          "ambiguous_match",
          "precondition_required",
          "internal_error",
          "backend_unavailable",
          "request_timeout"
        ]
      },
...
      "KbSlug": {
        "type": "string",
        "description": "A knowledge base slug: 1 to 40 of `a-z`, `0-9` and `-`, not starting or ending with `-`.",
        "pattern": "^[a-z0-9]([a-z0-9-]{0,38}[a-z0-9])?$"
      },
...
      "Refusal": {
        "type": "object",
        "description": "The error envelope of a request the server refused before handling it: at its limit of requests in flight, or out of time.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "`backend_unavailable` or `request_timeout`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request's id, when one had been assigned."
          }
        }
      },
...
    },
    "responses": {
      "BackendUnavailable": {
        "description": "`backend_unavailable`: a backend or queue is unavailable or full, or the server is at its limit of requests in flight. Retry after `Retry-After`. A write answered this way may already be stored; repeat it.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            },
            "description": "Seconds to wait: `5`."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "$ref": "#/components/schemas/Refusal"
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "The request is malformed: `invalid_request` or `malformed_range`. A query string the server cannot parse at all is answered `400` with a plain-text body instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "`forbidden`: the credential is valid and the knowledge base's access policy does not grant this operation. An anonymous caller is answered `404` instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Gone": {
        "description": "`gone`: the events after `Last-Event-ID` are no longer retained. Resubscribe without it.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "`not_found`: no such knowledge base or object — or one an anonymous caller may not see, which reads identically.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "`unauthorized`: no credential where one is required, or one that did not verify.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "`Bearer resource_metadata=\"…\"`, when the deployment publishes OAuth protected resource metadata."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "service token or OIDC access token (JWT)",
        "description": "`Authorization: Bearer <token>`: the deployment's service token, or an access token from the configured OIDC issuer. A malformed or unverified credential is always `401`, never treated as anonymous."
      }
    }
  },
...
}